+++ src/oglwidget.h
@@ -29,7 +29,7 @@ public:
     QVector3D screenToWorld(QPoint pos);
     void loadData(USV::CaseData &case_data);
     void updatePositions(const std::vector<USV::Vessel>& vessels);
-    void updateTime(float t);
+    void updateTime(double t);
     void wheelEvent ( QWheelEvent * event );
     void mousePressEvent(QMouseEvent* event);
     void mouseReleaseEvent(QMouseEvent*event);
@@ -70,7 +70,7 @@ protected:
     GLGrid *grid;
     GLSea *sea;
     GLRestrictions* restrictions;
-    float time{0.0f};
+    double time{0.0f};
     std::unique_ptr<USV::CaseData> case_data_;
 public:
     [[nodiscard]] const USV::CaseData* case_data() const {
